+++ src/Tree.ts
@@ -21,7 +21,7 @@ import { Comonad1 } from './Comonad'
 import { Eq, fromEquals } from './Eq'
 import { Extend1 } from './Extend'
 import { Foldable1 } from './Foldable'
-import { identity, pipe } from './function'
+import { flow, identity, pipe } from './function'
 import { bindTo as bindTo_, Functor1, tupled as tupled_ } from './Functor'
 import { HKT, Kind, Kind2, Kind3, Kind4, URIS, URIS2, URIS3, URIS4 } from './HKT'
 import {
@@ -103,35 +103,38 @@ export const unfoldForest = <B, A>(f: (b: B) => readonly [A, ReadonlyArray<B>]) 
  */
 export function unfoldTreeM<M extends URIS4>(
   M: Monad4<M> & Applicative4<M>
-): <B, S, R, E, A>(b: B, f: (b: B) => Kind4<M, S, R, E, readonly [A, ReadonlyArray<B>]>) => Kind4<M, S, R, E, Tree<A>>
+): <B, S, R, E, A>(
+  f: (b: B) => Kind4<M, S, R, E, readonly [A, ReadonlyArray<B>]>
+) => (b: B) => Kind4<M, S, R, E, Tree<A>>
 export function unfoldTreeM<M extends URIS3>(
   M: Monad3<M> & Applicative3<M>
-): <B, R, E, A>(b: B, f: (b: B) => Kind3<M, R, E, readonly [A, ReadonlyArray<B>]>) => Kind3<M, R, E, Tree<A>>
+): <B, R, E, A>(f: (b: B) => Kind3<M, R, E, readonly [A, ReadonlyArray<B>]>) => (b: B) => Kind3<M, R, E, Tree<A>>
 export function unfoldTreeM<M extends URIS3, E>(
   M: Monad3C<M, E> & Applicative3C<M, E>
-): <B, R, A>(b: B, f: (b: B) => Kind3<M, R, E, readonly [A, ReadonlyArray<B>]>) => Kind3<M, R, E, Tree<A>>
+): <B, R, A>(f: (b: B) => Kind3<M, R, E, readonly [A, ReadonlyArray<B>]>) => (b: B) => Kind3<M, R, E, Tree<A>>
 export function unfoldTreeM<M extends URIS2>(
   M: Monad2<M> & Applicative2<M>
-): <B, E, A>(b: B, f: (b: B) => Kind2<M, E, readonly [A, ReadonlyArray<B>]>) => Kind2<M, E, Tree<A>>
+): <B, E, A>(f: (b: B) => Kind2<M, E, readonly [A, ReadonlyArray<B>]>) => (b: B) => Kind2<M, E, Tree<A>>
 export function unfoldTreeM<M extends URIS2, E>(
   M: Monad2C<M, E> & Applicative2C<M, E>
-): <B, A>(b: B, f: (b: B) => Kind2<M, E, readonly [A, ReadonlyArray<B>]>) => Kind2<M, E, Tree<A>>
+): <B, A>(f: (b: B) => Kind2<M, E, readonly [A, ReadonlyArray<B>]>) => (b: B) => Kind2<M, E, Tree<A>>
 export function unfoldTreeM<M extends URIS>(
   M: Monad1<M> & Applicative1<M>
-): <B, A>(b: B, f: (b: B) => Kind<M, readonly [A, ReadonlyArray<B>]>) => Kind<M, Tree<A>>
+): <B, A>(f: (b: B) => Kind<M, readonly [A, ReadonlyArray<B>]>) => (b: B) => Kind<M, Tree<A>>
 export function unfoldTreeM<M>(
   M: Monad_<M> & Applicative_<M>
-): <B, A>(b: B, f: (b: B) => HKT<M, readonly [A, ReadonlyArray<B>]>) => HKT<M, Tree<A>>
+): <B, A>(f: (b: B) => HKT<M, readonly [A, ReadonlyArray<B>]>) => (b: B) => HKT<M, Tree<A>>
 export function unfoldTreeM<M>(
   M: Monad_<M> & Applicative_<M>
-): <B, A>(b: B, f: (b: B) => HKT<M, readonly [A, ReadonlyArray<B>]>) => HKT<M, Tree<A>> {
+): <B, A>(f: (b: B) => HKT<M, readonly [A, ReadonlyArray<B>]>) => (b: B) => HKT<M, Tree<A>> {
   const unfoldForestMM = unfoldForestM(M)
-  return <A, B>(b: B, f: (b: B) => HKT<M, readonly [A, ReadonlyArray<B>]>) =>
-    pipe(
-      f(b),
+  return (f) =>
+    flow(
+      f,
       M.chain(([value, bs]) =>
         pipe(
-          unfoldForestMM(f)(bs),
+          bs,
+          unfoldForestMM(f),
           M.map((forest) => ({ value, forest }))
         )
       )
@@ -177,7 +180,7 @@ export function unfoldForestM<M>(
   M: Monad_<M> & Applicative_<M>
 ): <B, A>(f: (b: B) => HKT<M, readonly [A, ReadonlyArray<B>]>) => (bs: ReadonlyArray<B>) => HKT<M, Forest<A>> {
   const traverseM = RA.traverse(M)
-  return (f) => traverseM((b) => unfoldTreeM(M)(b, f))
+  return (f) => traverseM(unfoldTreeM(M)(f))
 }
 
 // -------------------------------------------------------------------------------------