+++ test/Tree.ts
@@ -116,15 +116,24 @@ describe('Tree', () => {
 
   it('unfoldTreeM', async () => {
     U.deepStrictEqual(
-      _.unfoldTreeM({ ...O.Monad, ...O.Applicative })(1, (b) => O.some([b, b < 3 ? [b + 1, b + 2] : []])),
+      pipe(
+        1,
+        _.unfoldTreeM({ ...O.Monad, ...O.Applicative })((b) => O.some([b, b < 3 ? [b + 1, b + 2] : []]))
+      ),
       O.some(_.make(1, [_.make(2, [_.make(3), _.make(4)]), _.make(3)]))
     )
     U.deepStrictEqual(
-      _.unfoldTreeM({ ...O.Monad, ...O.Applicative })(1, (b) => (b < 3 ? O.some([b, [b + 1, b + 2]]) : O.none)),
+      pipe(
+        1,
+        _.unfoldTreeM({ ...O.Monad, ...O.Applicative })((b) => (b < 3 ? O.some([b, [b + 1, b + 2]]) : O.none))
+      ),
       O.none
     )
     U.deepStrictEqual(
-      await _.unfoldTreeM({ ...T.Monad, ...T.ApplicativePar })(1, (b) => T.of([b, b < 3 ? [b + 1, b + 2] : []]))(),
+      await pipe(
+        1,
+        _.unfoldTreeM({ ...T.Monad, ...T.ApplicativePar })((b) => T.of([b, b < 3 ? [b + 1, b + 2] : []]))
+      )(),
       _.make(1, [_.make(2, [_.make(3), _.make(4)]), _.make(3)])
     )
   })